each function
for… in should work on arrays - it does in node, but browsers and/or
chai seem to break it - requiring the switch statement. This could be
extended to handle any iterable pretty easy this way, but not
generically. to do that I’d just check for iterable (array edge case)
and split the space into those two possibilities